Add loc for literal and refactor a bit
