+++ src/Juvix/Prelude.hs
@@ -3,6 +3,7 @@ module Juvix.Prelude
     module Juvix.Prelude.Error,
     module Juvix.Prelude.Files,
     module Juvix.Prelude.Lens,
+    module Juvix.Prelude.Loc,
   )
 where
 
@@ -10,3 +11,4 @@ import Juvix.Prelude.Base
 import Juvix.Prelude.Error
 import Juvix.Prelude.Files
 import Juvix.Prelude.Lens
+import Juvix.Prelude.Loc
