+++ src/Juvix/Prelude/Loc.hs
@@ -1,4 +1,4 @@
-module Juvix.Syntax.Concrete.Loc where
+module Juvix.Prelude.Loc where
 
 import Juvix.Prelude.Base
 import Prettyprinter
@@ -71,29 +71,11 @@ getLocSpan = foldr1 (<>) . fmap getLoc
 instance Semigroup Interval where
   Interval f s e <> Interval _f s' e' = Interval f (min s s') (max e e')
 
-data WithLoc a = WithLoc
-  { _withLocInt :: Interval,
-    _withLocParam :: a
-  }
-  deriving stock (Show)
 
 makeLenses ''Interval
 makeLenses ''FileLoc
 makeLenses ''Loc
 makeLenses ''Pos
-makeLenses ''WithLoc
-
-instance HasLoc (WithLoc a) where
-  getLoc = (^. withLocInt)
-
-instance Eq a => Eq (WithLoc a) where
-  (==) = (==) `on` (^. withLocParam)
-
-instance Ord a => Ord (WithLoc a) where
-  compare = compare `on` (^. withLocParam)
-
-instance Functor WithLoc where
-  fmap = over withLocParam
 
 singletonInterval :: Loc -> Interval
 singletonInterval l =
