+++ src/Juvix/Syntax/Concrete/Lexer.hs
@@ -4,6 +4,7 @@ import Data.Text qualified as Text
 import GHC.Unicode
 import Juvix.Internal.Strings qualified as Str
 import Juvix.Prelude
+import Juvix.Syntax.Loc
 import Juvix.Syntax.Concrete.Base hiding (Pos, space)
 import Juvix.Syntax.Concrete.Base qualified as P
 import Juvix.Syntax.Concrete.Parser.InfoTableBuilder
