+++ src/Juvix/Syntax/Concrete/Parser.hs
@@ -262,12 +262,12 @@ hole = snd <$> interval kwHole
 literalInteger :: Members '[Reader ParserParams, InfoTableBuilder] r => ParsecS r LiteralLoc
 literalInteger = do
   (x, loc) <- integer
-  return (LiteralLoc (LitInteger x) loc)
+  return (WithLoc loc (LitInteger x) )
 
 literalString :: Members '[Reader ParserParams, InfoTableBuilder] r => ParsecS r LiteralLoc
 literalString = do
   (x, loc) <- string
-  return (LiteralLoc (LitString x) loc)
+  return (WithLoc loc (LitString x))
 
 literal :: Members '[Reader ParserParams, InfoTableBuilder] r => ParsecS r LiteralLoc
 literal = do
