+++ src/Juvix/Syntax/Concrete/Parser/InfoTableBuilder.hs
@@ -11,7 +11,8 @@ module Juvix.Syntax.Concrete.Parser.InfoTableBuilder
 where
 
 import Juvix.Prelude
-import Juvix.Syntax.Concrete.LiteralLoc
+import Juvix.Syntax.Concrete.Literal
+import Juvix.Syntax.Loc
 import Juvix.Syntax.Concrete.Parser.InfoTable
 import Juvix.Syntax.Concrete.Parser.ParsedItem
 
@@ -37,6 +38,8 @@ registerKeyword i =
         _parsedTag = ParsedTagKeyword
       }
 
+
+
 registerLiteral :: Member InfoTableBuilder r => LiteralLoc -> Sem r LiteralLoc
 registerLiteral l =
   l
@@ -46,7 +49,7 @@ registerLiteral l =
           _parsedTag = tag
         }
   where
-    tag = case l ^. literalLocLiteral of
+    tag = case l ^. withLocParam of
       LitString {} -> ParsedTagLiteralString
       LitInteger {} -> ParsedTagLiteralInt
     loc = getLoc l
