+++ src/Juvix/Syntax/Concrete/Scoped/Pretty/Base.hs
@@ -702,9 +702,6 @@ instance PrettyCode Literal where
     LitInteger n -> return $ annotate AnnLiteralInteger (pretty n)
     LitString s -> return $ ppStringLit s
 
-instance PrettyCode LiteralLoc where
-  ppCode l = ppCode (l ^. literalLocLiteral)
-
 instance PrettyCode AxiomRef where
   ppCode a = ppCode (a ^. axiomRefName)
